+++ server/controllers/api/channelClaims.js
@@ -1,4 +1,4 @@
-const { getChannelClaims } = require('../../controllers/serveController.js');
+const { getChannelClaims } = require('../utils/serveHelpers.js');
 const { handleErrorResponse } = require('../utils/errorHandlers.js');
 
 const NO_CHANNEL = 'NO_CHANNEL';
