+++ server/controllers/api/claimLongId.js
@@ -1,4 +1,4 @@
-const { getClaimId } = require('../../controllers/serveController.js');
+const { getClaimId } = require('../utils/serveHelpers.js');
 const { handleErrorResponse } = require('../utils/errorHandlers.js');
 
 const NO_CHANNEL = 'NO_CHANNEL';
