+++ server/controllers/assets/serveAssetByClaim.js
