+++ server/controllers/assets/serveAssetByIdentifierAndClaim.js
