+++ server/controllers/auth/logout.js
