+++ server/controllers/pages/redirect.js
