+++ server/routes/api/index.js
@@ -1,16 +1,16 @@
-const channelAvailability = require('./channelAvailability');
-const channelClaims = require('./channelClaims');
-const channelData = require('./channelData');
-const channelShortId = require('./channelShortId');
-const claimAvailability = require('./claimAvailability');
-const claimData = require('./claimData');
-const claimGet = require('./claimGet');
-const claimLongId = require('./claimLongId');
-const claimPublish = require('./claimPublish');
-const claimResolve = require('./claimResolve');
-const claimShortId = require('./claimShortId');
-const claimList = require('./claimList');
-const fileAvailability = require('./fileAvailability');
+const channelAvailability = require('../../controllers/api/channelAvailability');
+const channelClaims = require('../../controllers/api/channelClaims');
+const channelData = require('../../controllers/api/channelData');
+const channelShortId = require('../../controllers/api/channelShortId');
+const claimAvailability = require('../../controllers/api/claimAvailability');
+const claimData = require('../../controllers/api/claimData');
+const claimGet = require('../../controllers/api/claimGet');
+const claimLongId = require('../../controllers/api/claimLongId');
+const claimPublish = require('../../controllers/api/claimPublish');
+const claimResolve = require('../../controllers/api/claimResolve');
+const claimShortId = require('../../controllers/api/claimShortId');
+const claimList = require('../../controllers/api/claimList');
+const fileAvailability = require('../../controllers/api/fileAvailability');
 
 const multipartMiddleware = require('../utils/multipartMiddleware');
