+++ server/routes/assets/index.js
@@ -1,5 +1,5 @@
-const serveAssetByClaim = require('./serveAssetByClaim');
-const serveAssetByIdentifierAndClaim = require('./serveAssetByIdentifierAndClaim');
+const serveAssetByClaim = require('../../controllers/assets/serveAssetByClaim');
+const serveAssetByIdentifierAndClaim = require('../../controllers/assets/serveAssetByIdentifierAndClaim');
 
 module.exports = (app, db) => {
   app.get('/:identifier/:claim', serveAssetByIdentifierAndClaim);
