+++ server/routes/fallback/index.js
@@ -1,4 +1,4 @@
-const handlePageRequest = require('./sendReactApp');
+const handlePageRequest = require('../../controllers/fallback/sendReactApp');
 
 module.exports = (app) => {
   app.get('*', handlePageRequest);
